Move dependency check to new job_utils
