+++ turbinia/task_utils.py
@@ -12,7 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Task runner for Turbinia."""
+"""Task utilities for Turbinia."""
 
 from datetime import datetime
 import logging
