+++ turbinia/worker.py
@@ -26,6 +26,7 @@ from turbinia import config
 from turbinia.config import logger
 from turbinia import task_utils
 from turbinia import TurbiniaException
+from turbinia import job_utils
 from turbinia.lib import docker_manager
 from turbinia.jobs import manager as job_manager
 from turbinia.tcelery import TurbiniaCelery
@@ -165,29 +166,6 @@ def check_directory(directory):
           'Can not add write permissions to {0:s}'.format(directory))
 
 
-def register_job_timeouts(dependencies):
-  """Registers a timeout for each job.
-
-  Args:
-    dependencies(dict): dependencies to grab timeout value from.
-  """
-  log.info('Registering job timeouts.')
-  timeout_default = 3600
-
-  job_names = list(job_manager.JobsManager.GetJobNames())
-  # Iterate through list of jobs
-  for job, values in dependencies.items():
-    if job not in job_names:
-      continue
-    timeout = values.get('timeout')
-    if not isinstance(timeout, int):
-      log.warning(
-          'No timeout found for job: {0:s}. Setting default timeout of {1:d} seconds.'
-          .format(job, timeout_default))
-      timeout = timeout_default
-    job_manager.JobsManager.RegisterTimeout(job, timeout)
-
-
 class TurbiniaWorkerBase:
   """Base class for Turibinia Workers."""
 
@@ -225,7 +203,7 @@ class TurbiniaWorkerBase:
     check_directory(config.MOUNT_DIR_PREFIX)
     check_directory(config.OUTPUT_DIR)
     check_directory(config.TMP_DIR)
-    register_job_timeouts(dependencies)
+    job_utils.register_job_timeouts(dependencies)
 
     jobs = job_manager.JobsManager.GetJobNames()
     log.info(
